renamed mini_lm, moved requirements from VDBS
