+++ ragna/embedding_models/__init__.py
@@ -2,4 +2,4 @@ __all__ = [
     "MiniLML6v2"
 ]
 
-from ._embedding import MiniLML6v2
+from ._minilm_l6_v2 import MiniLML6v2
